MCH: some refactoring (mainly moving things around)
This commit does not introduce any new feature, but just reorganize a
bit where some files are located. In particular :

- the TrackerParam is moved to MCHBase to be useable from e.g.
  TimeClustering w/o requiring TimeClustering to depend on MCHTracking
- the ROF filtering used in TimeClustering (based on the trackable condition)
is moved to its own module (in preparation for the addition of a new ROF
filter soon)
