+++ Detectors/MUON/MCH/Base/src/MCHBaseLinkDef.h
@@ -18,4 +18,7 @@
 #pragma link C++ namespace o2;
 #pragma link C++ namespace o2::mch;
 
+#pragma link C++ class o2::mch::TrackerParam + ;
+#pragma link C++ class o2::conf::ConfigurableParamHelper < o2::mch::TrackerParam> + ;
+
 #endif
