+++ Detectors/MUON/MCH/README.md
@@ -14,6 +14,8 @@ This is a top page for the MCH detector documentation.
 \subpage refDetectorsMUONMCHContour
 \subpage refDetectorsMUONMCHDevIO
 \subpage refDetectorsMUONMCHDigitFiltering
+\subpage refDetectorsMUONMCHROFFiltering
+\subpage refDetectorsMUONMCHTimeClustering
 \subpage refDetectorsMUONMCHGeometry
 \subpage refDetectorsMUONMCHMapping
 \subpage refDetectorsMUONMCHRaw
