+++ Detectors/MUON/MCH/ROFFiltering/src/TrackableFilter.cxx
@@ -9,13 +9,4 @@
 // granted to it by virtue of its status as an Intergovernmental Organization
 // or submit itself to any jurisdiction.
 
-#ifdef __CLING__
-
-#pragma link off all globals;
-#pragma link off all classes;
-#pragma link off all functions;
-
-#pragma link C++ class o2::mch::TrackerParam + ;
-#pragma link C++ class o2::conf::ConfigurableParamHelper < o2::mch::TrackerParam> + ;
-
-#endif
+#include "MCHROFFiltering/TrackableFilter.h"
