+++ Detectors/MUON/MCH/Tracking/CMakeLists.txt
@@ -17,8 +17,4 @@ o2_add_library(MCHTracking
         src/TrackFitter.cxx
         src/TrackFinderOriginal.cxx
         src/TrackFinder.cxx
-        src/TrackerParam.cxx
         PUBLIC_LINK_LIBRARIES O2::Field O2::MCHBase O2::Framework O2::CommonUtils)
-
-o2_target_root_dictionary(MCHTracking
-                          HEADERS include/MCHTracking/TrackerParam.h)
