+++ Detectors/MUON/MCH/Tracking/src/TrackFinderOriginal.cxx
@@ -24,8 +24,8 @@
 #include <TMath.h>
 
 #include "Field/MagneticField.h"
+#include "MCHBase/TrackerParam.h"
 #include "MCHTracking/TrackExtrap.h"
-#include "MCHTracking/TrackerParam.h"
 
 namespace o2
 {
