+++ Detectors/MUON/MCH/Workflow/CMakeLists.txt
@@ -10,11 +10,6 @@
 # or submit itself to any jurisdiction.
 
 
-add_library(obj-mch-reco OBJECT)
-target_sources(obj-mch-reco PRIVATE src/TimeClusterFinderSpec.cxx)
-target_link_libraries(obj-mch-reco PUBLIC O2::Framework O2::MCHTimeClustering
- O2::MCHTracking)
-
 # MCHWorkflow library is (at least) needed by Detectors/CTF/workflow
 o2_add_library(MCHWorkflow
                SOURCES
@@ -82,13 +77,6 @@ o2_add_executable(
         COMPONENT_NAME mch
         PUBLIC_LINK_LIBRARIES O2::Framework O2::MCHTriggering)
 
-o2_add_executable(
-        digits-to-timeclusters-workflow
-        SOURCES src/digits-to-timeclusters-workflow.cxx
-        COMPONENT_NAME mch
-        PUBLIC_LINK_LIBRARIES O2::MCHBase
-        TARGETVARNAME mch-time-cluster-finder)
-target_link_libraries(${mch-time-cluster-finder} PRIVATE obj-mch-reco)
 
 o2_add_executable(
         digits-to-preclusters-workflow
@@ -237,13 +225,13 @@ o2_add_executable(
             O2::MCHDigitFiltering
             O2::MCHGeometryTransformer
             O2::MCHMappingImpl4
+            O2::MCHTimeClustering
             O2::MCHTriggering
             O2::MCHTracking
             O2::MCHWorkflow
             O2::SimulationDataFormat
             O2::Steer
         TARGETVARNAME mch-reco-workflow)
-target_link_libraries(${mch-reco-workflow} PRIVATE obj-mch-reco)
 
 o2_add_executable(tracks-file-dumper
         SOURCES src/tracks-file-dumper.cxx
