+++ Detectors/MUON/MCH/Workflow/src/reco-workflow.cxx
@@ -21,10 +21,10 @@
 #include "Framework/Variant.h"
 #include "Framework/WorkflowSpec.h"
 #include "MCHDigitFiltering/DigitFilteringSpec.h"
+#include "MCHTimeClustering/TimeClusterFinderSpec.h"
 #include "MCHWorkflow/ClusterFinderOriginalSpec.h"
 #include "MCHWorkflow/PreClusterFinderSpec.h"
 #include "MCHWorkflow/TrackWriterSpec.h"
-#include "TimeClusterFinderSpec.h"
 #include "TrackFinderSpec.h"
 #include "TrackFitterSpec.h"
 #include "TrackMCLabelFinderSpec.h"
